# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: IRIS Enterprise Commerce API
  version: 0.1.0
  description: Source-of-truth contract owned by the backend team.
servers:
  - url: /api
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: User profile and linked identities.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/identities:
    get:
      operationId: getLinkedIdentities
      tags: [Auth]
      description: >-
        IRIS ID linked-providers view for the Account Center. Reports every provider (discord,
        steam, epic); unlinked providers appear with null fields and `canUnlink: false`.
        `canUnlink` enforces the invariant that at least one provider must remain linked.
        Mirrors `contracts/fixtures/linked-identities.json`.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Linked identities and IRIS ID linking rules.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkedIdentitiesResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/link-steam:
    post:
      operationId: linkSteam
      tags: [Auth]
      description: >-
        Link a Steam identity to the authenticated IRIS ID. Linking requires
        proof-of-control of the provider (Steam OpenID assertion verified server-side or via
        the provider callback). IRIS ID NEVER auto-merges accounts by email or display name.
        `proof` is optional in development (the verified callback cookie is used) but REQUIRED
        in production.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [steamId]
              properties:
                steamId: { type: string }
                proof: { $ref: '#/components/schemas/LinkProof' }
      responses:
        '200':
          description: Steam account linked successfully.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/unlink-steam:
    post:
      operationId: unlinkSteam
      tags: [Auth]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Steam account unlinked successfully.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/link-epic:
    post:
      operationId: linkEpic
      tags: [Auth]
      description: >-
        Link an Epic Games identity to the authenticated IRIS ID. Requires proof-of-control
        of the provider (Epic OAuth verified server-side). IRIS ID NEVER auto-merges accounts
        by email or display name. `proof` is optional in development but REQUIRED in production.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [epicId]
              properties:
                epicId: { type: string }
                proof: { $ref: '#/components/schemas/LinkProof' }
      responses:
        '200':
          description: Epic Games account linked successfully.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/unlink-epic:
    post:
      operationId: unlinkEpic
      tags: [Auth]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Epic Games account unlinked successfully.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/unlink-discord:
    post:
      operationId: unlinkDiscord
      tags: [Auth]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Discord account unlinked successfully.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions:
    get:
      operationId: listSessions
      tags: [Auth]
      description: >-
        Active device sessions for the user. Each item carries risk-based login flags derived
        from the requesting device: `isCurrent`, `riskFlag` (true on a new IP/device), and
        `riskReasons`. The frontend can surface a security prompt for any flagged session.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Active sessions for the user, with derived risk flags.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SessionWithRisk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [Auth]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: Session revoked successfully.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /auth/recovery:
    post:
      operationId: requestRecovery
      tags: [Auth]
      description: >-
        Start an account recovery. STUB: schema-complete, provider proof verification lands in
        a later milestone. Recovery is ALWAYS a proof-of-control challenge against a still-linked
        provider; accounts are never recovered or merged by email/display name. The endpoint is
        intentionally non-enumerating (it does not confirm whether an account exists).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecoveryRequest' }
      responses:
        '202':
          description: Recovery ticket accepted; proof-of-control required to continue.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecoveryTicket' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: token }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    legacyApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        DEPRECATED legacy plugin auth (per-user API key + X-Server-Id + IP whitelist). Accepted
        ONLY during the CR-PLUGIN-006 migration overlap window so plugins already in the field do
        not break; each use is logged as a deprecation. New deployments MUST use `hmacAuth`. Will
        be removed once all servers have migrated to signed requests.
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: >-
        Signed plugin request authentication (HMAC-SHA256). The signature is carried in
        `X-Signature` and is computed over a canonical string with the per-credential HMAC
        secret. The signing IDENTITY and the signing SECRET are two distinct values:


        Headers (all required):

        - `X-Plugin-Key-Id`: the credential identifier. It IDENTIFIES which signing secret to
          use; it is NOT itself the secret and is safe to log. It is rotatable and revocable
          independently of the secret.

        - `X-Plugin-Version`: plugin build/version string (compatibility + telemetry).

        - `X-Request-Timestamp`: unix epoch milliseconds. Requests outside the clock-skew
          window (+/- 5 minutes / 300000 ms) are rejected.

        - `X-Request-Nonce`: >= 16 chars, single-use; replays within the skew window are
          rejected (nonce cache).

        - `X-Content-Sha256`: hex SHA-256 of the exact JSON request body ('' when no body).

        - `X-Signature`: `HMAC_SHA256(secret, canonicalString)` as lowercase hex, where
          `canonicalString = METHOD \n PATH_AND_QUERY \n TIMESTAMP \n NONCE \n CONTENT_SHA256`.

...
  responses:
    Unauthorized:
      description: Authentication required
    BadRequest:
      description: Invalid request parameters or state conflict
...
    NotFound:
      description: Resource not found
  schemas:
    User:
      type: object
      required: [id, discordId, pointsBalance, role, createdAt]
      properties:
        id: { type: string }
        steamId: { type: [string, 'null'] }
        epicId: { type: [string, 'null'] }
        discordId: { type: string }
        discordUsername: { type: [string, 'null'] }
        discordAvatar: { type: [string, 'null'] }
        pointsBalance: { type: string }
        role: { type: string }
        createdAt: { type: string, format: date-time }
...
    SessionWithRisk:
      allOf:
        - { $ref: '#/components/schemas/UserSession' }
        - type: object
          required: [isCurrent, riskFlag, riskReasons]
          properties:
            isCurrent:
              type: boolean
              description: True when this session matches the requesting device's IP and user-agent.
            riskFlag:
              type: boolean
              description: Risk-based login flag — true when a non-current session is on a new IP or new device.
            riskReasons:
              type: array
              items: { type: string, enum: [new_ip, new_device] }
    LinkProof:
      type: object
      description: >-
        Proof-of-control material for account linking. `method` declares how control was
        proven. email/name/username are explicitly rejected — IRIS ID never merges by identity
        attributes. Method-specific verified fields (e.g. claimedId, verified token reference)
        may accompany `method` as additional properties.
      required: [method]
      properties:
        method:
          type: string
          enum: [steam_openid, epic_oauth, discord_oauth, provider_callback]
      additionalProperties: true
    LinkedIdentity:
      type: object
      required: [provider, providerAccountId, linkedAt, proofMethod, isPrimary, canUnlink]
      properties:
        provider: { type: string, enum: [discord, steam, epic] }
        providerAccountId: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        linkedAt: { type: [string, 'null'], format: date-time }
        proofMethod: { type: [string, 'null'] }
        isPrimary: { type: boolean }
        canUnlink:
          type: boolean
          description: >-
            True only when the provider is linked AND unlinking it would still leave at least one
            linked provider (minimumLinkedProviders invariant).
    LinkedIdentitiesResponse:
      type: object
      required: [userId, identities, rules]
      properties:
        userId: { type: string }
        identities:
          type: array
          items: { $ref: '#/components/schemas/LinkedIdentity' }
        rules:
          type: object
          required: [autoMergeByEmailOrName, proofOfControlRequired, minimumLinkedProviders]
          properties:
            autoMergeByEmailOrName: { type: boolean, const: false }
            proofOfControlRequired: { type: boolean, const: true }
            minimumLinkedProviders: { type: integer, const: 1 }
            note: { type: string }
    RecoveryRequest:
      type: object
      required: [provider, providerAccountId]
      properties:
        provider: { type: string, enum: [discord, steam, epic] }
        providerAccountId:
          type: string
          description: The provider-side account id the user claims; existence is never confirmed in the response.
        contactChannel:
          type: [string, 'null']
          description: Optional channel hint for delivering the challenge; never used to auto-merge accounts.
    RecoveryTicket:
      type: object
      required: [status, recoveryTicketId, provider, challengeNonce, expiresAt]
      properties:
        status: { type: string, const: pending_proof_of_control }
        recoveryTicketId: { type: string }
        provider: { type: string, enum: [discord, steam, epic] }
        challengeNonce: { type: string }
        instructions: { type: string }
        expiresAt: { type: string, format: date-time }
...